Update gui.py
Wrote a hard code version of the window that takes the contents of data entries and assigns them to each variable name. The variables are global, and so should be easily called to another file when this one is imported. There are no checks to see if submitted answers are usable, though.